refactor: Implements middleware to compare api deployed Date with the request Date

+++ helper-function/populate-database.ts
@@ -1,12 +1,12 @@
-import connectToDatabase from '../models/connectToDatabase';
 import { scrapeFights, scrapeFightsCard } from './scraper';
 import * as FightService from '../services/fight-service';
 import * as FightsCardService from '../services/fights-card-service';
 
-const populateDatabase = async () => {
-  connectToDatabase();
-  await Promise.all([FightService.deleteMany(), FightsCardService.deleteMany()]);
+const populateDatabase = async (deployedDate: string) => {
+  if (new Date().toLocaleDateString() <= deployedDate) return;
   
+  await Promise.all([FightService.deleteMany(), FightsCardService.deleteMany()]);
+
   const fights = await scrapeFights();
 
   await Promise.all(
@@ -15,9 +15,6 @@ const populateDatabase = async () => {
       scrapeFightsCard(fights),
     ],
   ).then((values) => FightsCardService.create(values[1]));
-  process.exit(0);
 };
-
-populateDatabase();
 
 export default populateDatabase;
